Add _y_spt_afterframes_await_load (#12)
* Add _y_spt_afterframes_await_load

NOTE: Totally alpha, unresearched and with NO guarantee of accuracy
regarding timing and stuff.

This would mainly be used for playing back a TAS with multiple parts
to save the user from having to bind keys to hundreds of scripts.

* Fix a style thing and remove a pointless comment.

Can't get rid of the extra tab character for some reason. :-/